Outsourced CoolDates component
