+++ src/components/Home.js
@@ -3,8 +3,7 @@ import Typography from "@material-ui/core/Typography";
 import CssBaseline from "@material-ui/core/CssBaseline";
 import Container from "@material-ui/core/Container";
 import { makeStyles } from "@material-ui/core/styles";
-import EventBusyIcon from "@material-ui/icons/EventBusy";
-import EventAvailableIcon from "@material-ui/icons/EventAvailable";
+import CoolDates from "./subcomponents/CoolDates";
 import Grid from "@material-ui/core/Grid";
 import Card from "@material-ui/core/Card";
 import CardContent from "@material-ui/core/CardContent";
@@ -40,24 +39,7 @@ const FixedCard = ({ id, title, values }) => {
     </Card>
   );
 };
-const CoolDates = ({ status, time }) => {
-  //  console.log(Icon);
-  return (
-    <div style={{ position: "relative" }}>
-      {status === "startDate" ? (
-        <EventAvailableIcon color="primary" />
-      ) : (
-        <EventBusyIcon color="error" />
-      )}
 
-      <span
-        style={{ marginLeft: "1rem", position: "absolute", bottom: "0.35rem" }}
-      >
-        {time ? time.substring(0, 10).replace(/-/g, "/") : null}
-      </span>
-    </div>
-  );
-};
 export const Home = ({
   from,
   to,
